[frontend] Extract dialog expectation
